tdf_media_handler/requirements.txt
obs-websocket-py
pyautogui
notifypy
flask
pywin32
pywinauto